+++ NEWS.md
@@ -4,7 +4,7 @@ Version 0.1.4.999: February 7, 2017
 General:
 
 + License changed to Creative Commons Attribution-ShareAlike 4.0 International
-(CC BY-SA 4.0).
+  (CC BY-SA 4.0).
 
 Selection Analysis: